requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
inky==1.5.0
numpy==2.0.1
pillow==10.4.0
requests==2.32.3
RPi.GPIO==0.7.1
smbus2==0.4.3
spidev==3.6
urllib3==2.2.2
